octopusEnergy/octopusEnergyFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pg8000
azure-identity
azure-keyvault-secrets